requirements.txt
ansiblemetrics==0.3.12
PyGithub==1.55
repository-miner==0.9.6
requests==2.26.0
tosca-metrics==0.0.11